fix: use real time in RPC middleware tests
The tests occasionally fail. The likely reason is that tokio simulated
time (used for testing) does not play nicely with actions that can take
real time. An example of such an action is a HTTP request to an RPC
server. While both the client and the server both live in the same
process, the request is still routed through syscalls. The suspension of
a task by syscall makes tokio advance time before it was intended.

This changes the tests to use wall clock time instead. The timeouts have
been shortened to keep the test running time at a reasonable level. No
test should take any more than just over 3 seconds. While it's not much,
it will still limit the amount of parallelism available to other testing
tasks to some degree.