build: remove the `sp1` tool installation step from CI
The tests no longer require the `sp1` toolchain to be installed, relying
on the committed ELF binary only. There is a separate CI job to check
the committed ELF corresponds to the pessimistic proof sources.

This will make the CI job definitions a bit cleaner and hopefully they
will run a bit faster too.

* tools not required since #494