refactor: clean up contract generics (#914)
A small clean up of contract and settlement generics, consisting of two
changes:

* Remove unused associated types in contract traits
* Make `submit_certificate_settlement` and `wait_for_settlement` into
standalone methods (called from the original trait implementation). That
allows for relaxing some generic bounds on the type parameters,
hopefully allowing for simpler testing / mocking setup when only one of
these methods needs to be tested.

## PR Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [ ] I have added or updated tests that comprehensively prove my change
is effective or that my feature works