chore: remove `serde` requirement from storage codec (#1033)
With the gradual move towards protobuf in storage, the requirement for
all storage types to implement the `serde` traits is unnecessary and a
bit limiting. This removes the `serde` supertraits from
`agglayer_storage::columns::Codec`. Consequently, we do not have
convenient default implementation for the methods so a macro is
introduced for it.