feat: database migration mechanism core function (#1211)
The first version of the database migration framework. It follows the
design described in agglayer/protocol-research#206. The user-facing core
functionality is ready. It allows one to use a database builder to
define a number of migration steps, starting at the "original" version
of the database.

See the tests for example usage for now. Proper docs will come later.

The migration integrity feature is largely missing. The idea is to take
a hash of the schema at each migration step and persist it in the
migration record column, together with some other metadata. Then verify
it during db initialization to check the migration steps have not been
tampered with. This is planned as a follow-up.

* design agglayer/protocol-research#206
* fixes #1212 
* related #1185